+++ presentation/src/main/java/oil/oing/nearby/presentation/ui/OilOingMain.kt
@@ -1,0 +1,4 @@
+package oil.oing.nearby.presentation.ui
+
+class OilOingMain {
+}
