+++ presentation/src/main/res/values/colors.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <!-- default colors-->
-    <color name="colorPrimary">#D53A47</color>
-    <color name="colorPrimaryDark">#D53A47</color>
-    <color name="colorAccent">#D53A47</color>
+    <color name="colorPrimary">#FF7A00</color>
+    <color name="colorPrimaryDark">#FF7A00</color>
+    <color name="colorAccent">#FF7A00</color>
     <color name="white_00">#FFFFFF</color>
     <color name="black_00">#000000</color>
     <color name="orange_00">#FFA726</color>
